Add search control to services page
Add a search bar to narrow down the services that are displayed on the
service-list page.
